requirements.txt
langchain
weaviate-client
openai